requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blinker=1.8.2=pypi_0
ca-certificates=2024.9.24=haa95532_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
filelock=3.13.1=pypi_0
flask=3.0.3=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.2.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
kiwisolver=1.4.7=pypi_0
libffi=3.4.4=hd77b12b_1
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.0=pypi_0
numpy=1.24.1=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pandas=2.0.3=pypi_0
pillow=10.2.0=pypi_0
pip=24.2=py38haa95532_0
psutil=6.1.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.20=h8205438_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
requests=2.32.3=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py38haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
# torch cuda는 컴퓨터에 맞게
torch=2.4.1+cu124=pypi_0
torchaudio=2.4.1+cu124=pypi_0
torchvision=0.19.1+cu124=pypi_0
# torch cuda는 컴퓨터에 맞게
tqdm=4.67.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.2=pypi_0
ultralytics=8.3.31=pypi_0
ultralytics-thop=2.0.11=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py38haa95532_0
zipp=3.20.2=pypi_0
